Change old to float and equalize test

+++ cppApplicationOld/src/shapes/mesh.cpp
@@ -2,13 +2,13 @@
 
 Mesh::Mesh(const std::vector<Triangle>& tris) : triangles(tris) {}
 
-std::optional<std::tuple<double, Eigen::Vector3d, const Triangle*>> Mesh::intersects(const Ray& ray) const {
-    std::optional<std::tuple<double, Eigen::Vector3d, const Triangle*>> nearest_hit;
-    double min_t = std::numeric_limits<double>::infinity();
+std::optional<std::tuple<float, Eigen::Vector3f, const Triangle*>> Mesh::intersects(const Ray& ray) const {
+    std::optional<std::tuple<float, Eigen::Vector3f, const Triangle*>> nearest_hit;
+    float min_t = std::numeric_limits<float>::infinity();
     for (const auto& triangle : triangles) {
         auto hit = triangle.intersects(ray);
         if (hit.has_value()) {
-            double t = std::get<0>(hit.value());
+            float t = std::get<0>(hit.value());
             if (t < min_t) {
                 min_t = t;
                 nearest_hit = std::make_tuple(t, std::get<1>(hit.value()), &triangle);
